pipeline: need to resolve the package's prefix with --no-cache as well
Because we're dealing with partial installations (where a lot of
packages are not present), the package we're building might be
"not selected" because it is not selected by the manifest.

Example: 'base/orogen/std' depends on 'rtt', the manifest
lists base/orogen/std

When building 'rtt', base/orogen/std is not present on disk
and therefore autoproj cannot know that 'rtt' is selected. Which
means that 'rtt' is *not* selected from autoproj's point of view,
and the installation manifest does not include it.